Gavin wanted an override of the year used for transient emissions, and wanted me to use an existing parameter to do it. So, with this commit all chem tracers' 2D surface sources and 3D aircraft source and (currently not used) GFED 3D biomass burning transient sources will use and repeat year abs(o3_yr) and ignore the current model year if o3_yr is non-zero.  If o3_yr=0 the transient sources use the model jyear.
So, a lot or replacing jyear-->xyear and jday-->xday in
TRACERS_SPECIAL_Shindell.f where xyear is passes as either jyear
or abs(o3_yr).

I have tested multiple combinations for the model year and o3_yr and the
correct emissions seem to be picked in each case, but of course let
me know if you have any problems. I will monitor the emissions totals
in the AR5 runs.

Also, in TRACERS_SPECIAL_Shindell.f: for the NOx aircraft source, I
hardcoded the routine not to read the aircraft file if xyear < 1900.
There remains the "bug" of excessive reading of this source, so this
will speed things up a bit.