WARNING: the ocean state now contains new prognostic arrays; after applying this update, you cannot continue previous runs without adding the new arrays to a restart file.
The alternate code in OCNDYN2.f is now the default; subroutine
OCEANS in OCNDYN.f has been renamed OCEANS_old.

OCNDYN2 and OCNKPP changes:

1. Semi-prognostic D-grid velocities have been introduced for the
computation of the coriolis force (the so-called "CD" scheme).

2. For simplicity with the C+D-grid velocities and to reduce CPU
time, the application of the KPP mixing scheme on quarter boxes has
been eliminated; this does not seem to affect results at 1- and
2-degree resolution (4-degree resolution hasn't been tested).

3. Vertical advection now uses a simple centered-difference (but
slightly upstream-biased) scheme in place of the LUS.  It is less
diffusive than the LUS with respect to vertical velocities that
oscillate in time.
#define LUS_VERT_ADV reinstates the LUS.

4. Subroutine ODIFF is only called every 3 hours (reduces CPU time).