Added new functionality: now one can automatically download all necessary input files from central repository (data portal). If one specifies "-w" flag input files will be downloaded from "data portal" repository (instead of being copied from local repository. The most typical usage of this functionality will be to update local repoitory of input files. For exmple, the command:
 get_input_data -w <my_rundeck> <GCMSEARCHPATH>

where <my_rundeck> is the name of a rundeck and <GCMSEARCHPATH> is
a path to local repository of input files as specified by  GCMSEARCHPATH
in ~.modelErc, will download all missing files necessary to run
<my_rundeck> and put them to GCMSEARCHPATH .

This functionality is only of interest to those who run the model
on their own machines (rather than on discover). The central repository
is still being tested and is not yet fully populated. You may want
to wait until further notice before using this.