There are three new rundeck parameters: prefDustSources, fracClayPDFscheme, fracSiltPDFscheme
prefDustSources serves to choose the appropriate emission parameters
for the files with the preferred dust sources. The files have to be
assigned to DSRC in the rundeck, accordingly.

For the 72x46 resolution, only one file is available for now:

DSRC=Ginoux2001_source_VegMask_72x46      for prefDustSources=0

For prefDustSources=0 ... 4, the emission parameters fracClayPDFscheme
and fracSiltPDFscheme are chosen by the model and any settings for
these two parameters in the rundeck are w/o any effect.

prefDustSources>=5 allows to set fracClayPDFscheme and
fracSiltPDFscheme in the rundeck.

Some minor dust aerosol related additions and changes to the rundecks
were made as well. However, they don't change any previous settings,
except for T4dusM20.R

rad_interact_aer=1 and Ndisk=1920 in E4TdusM20.R