Cleaning up some namespace issues associated with GEOM_B.
1) Some entities from CONSTANTS were being imported through GEOM,
which can be an issue at times.  Mostly this induced trivial changes
on just 3 other files.

2) Formally declared module to be default private and then added
"public" to just those entities that are needed by other modules.
(Most of them in this case actually, but I expect the list to go down
soon. Hence the exercise.)

3) Eliminated some unreferenced uses of variables which have no
representation on Cubed-sphere grid.