Adding UseRadAerInFastJ=1 and changing the SPECFJ file for photolysis. This goes along with code changes to chemistry/rad code whose commit notes were:
Jan, please confirm that in RAD_DRV and RAD_COM your needed ttausv stuff is still OK
(now loops over all layers, not just 11, and all tracers, as we discussed).

Changes by Apostolos (and Greg) to improve the gas-aerosol coupling in the model:
now the aerosol tracers can directly influence the photolysis of gases (in the
Fast-J2 photolysis code) and thus their chemistry and concentrations through this path.
Currently this development works only if all the aerosol tracers are defined in radiation code,
excluding the very coarse silt4, which can be ON but is not mandatory.
Namely, the aerosols taken into account are: SO4, seasalt1, seasalt2, OCIA, BCIA, BCB,
NO3p, clay (4 tracers), silt1, silt2, silt3. If this full compliment of tracers is not
defined in the rad code, the photolysis code will ignore the aerosols (just like it did until now),
and use an offline climatology. (In fact, at the moment, if you try to use this with some other
configuration you will get a model stop.) This will be improved in the future to work, with
subsets of aerosol tracers allowed as well.

On/Off is controlled by dbparam: UseRadAerInFastJ=1. Current default is OFF (0),
though this will change once more flexibility is introduced. When off, only a simple
offline BC profile is used (as it always has been, just with updated optical properties).
We are committing the flag to ON in the E4TcadF40.R template deck.

By comparing our results with UseRadAerInFastJ=1 versus what we get with
UseRadAerInFastJ=0, we find that there are considerable changes in OH (and
subsequent changes in other gases) above some regions with high aerosol concentrations.

For more details on this development or on the results, please contact Apostolos:
avoulgarakis@giss.nasa.gov.

For AR5 interested parties: we have tested this a lot in short runs: longer simulations
are going now, just in case. Reproducibility tests were done a week or so ago with
this code.