Decided to place the regrid-to-latlon option into the generic scaling routine.  This eliminates the need for temporary latlon acc-files and the complicated script which generates them. At this point I am assuming that if regridding is needed, it is because modelE was run on the cubed sphere. Later I may generalize to allow regridding from one latlon grid to another, and perhaps from other grids (e.g. tripolar).
Regridding is requested by simply specifying the name of a
file with the regridding info on the command line, after
the name of the acc-array to be processed.
Different quantities can be regridded using different
methods by specifying the requested method in the parts
of the modelE code that declare metadata for each output
quantity.  I am still deciding on the conventions for this.
Interpolation-based methods will be imported soon.