Rmoving the parameter useRadAerInFastJ, to go along with code commits as noted below:
Some changes related to the aerosols and their interaction to the photolysis code.

a) Bug-fix: the way that 3-D aerosol optical depth was defined before in RAD_DRV.f (ntrix(n) indexing), clay was not saved properly, since only the last size-bin (4th) was taken into account (4 times). We now create a new array (ttausv_ntrace) which is used in the photolysis code (TRCHEM_fastj2.f), and is indexed by n=1,NTRACE, so that the above-mentioned problem will be avoided. We get four separate clays.

b) The rad-code aerosols - photolysis interaction has been made more flexible, so that now it does not require all the aerosol tracers to be turned on in the model in order to run. So, now, you will not get a model stop with other configurations. For example, one can have only dust tracers turned on, and still the photolysis code will use only dust, and zero the rest of the aerosol optical depths.

c) The option to use the 4th (very coarse) silt tracer (silt4) has been added. So, if this is turned on in the model, the photolysis code will use it.