Because of round-off errors in the computations in the photolysis code, Apostolos and I noticed that the order of the aerosol optical depths passed to it slightly changed the photolysis rates. (Or even passing the same order but summing DTAUX in the reverse order).
(After much trying to get rid of this another way): This commit
reduces the precision of the photolysis rates to hopefully make
the order not matter. Pretty sure photolysis rates in Nature
are not known to 10 decimal places anyway...

So this will change results, but we do not expect any
large changes.