This commit will not change results unless you #define PBL_USES_GCM_TENDENCIES
It is an attempt to increase the correlation of PBL temperatures
with layer-1 values when layer-1 turbulence is too weak to
keep them in sync (a problem over the Antarctic plateau during
the polar night).  Increases of minimum diffusivities and drag
coefficients also help but are not being committed here.

The existing PBL formulation assumes the temperature tendency from
processes other than PBL turbulence is equal to PBL winds multiplied
by horizontal temperature gradients estimated from vertical differences
of the pressure gradient force.  This approximation may not be realistic
to the extent that cloud and radiative processes are important, or
subsidence is warming the first model layer, or the pressure gradient
force does not take the form assumed in modelE (FV dynamics uses a
different formulation).

#define PBL_USES_GCM_TENDENCIES enables the use of layer-1 temperature
tendencies from the host model computed as the difference between those
occurring before the PBL is called and those saved after the ATURB
calculations from the previous timestep.  The reason for saving
the temperatures after ATURB rather than the PBL is that the PBL holds
layer-1 temperatures fixed as a "boundary condition" for its
sub-layers between the surface and the midpoint of layer 1.

I also added the corresponding code for PBL winds, but it is commented
out until I determine its impacts.