Should not change results even when turned on, and in fact does very little unless you set #define AUXILIARY_OX_RADF (which is NOT on in E4TcadF40.R), namely: (1) The size of array TNFST_o3ref is increased. (2) An #ifdef ACCMIP_LIKE_DIAGS section in RAD_DRV was moved inside a #ifdef TRACERS_SPECIAL_Shindell section.
If you turn AUXILIARY_OX_RADF on, RCOMPX will be called again and you'll get
radiative forcing diags for either ozone climatology (if clim_interact_chem=1) or
ozone tracer (if clim_interact_chem=0). The reason for this is that sometimes
when one runs without using the tracer in the final RCOMPX call, one still wants the
ozone tracer's radiative forcing. This ability was lost in reorganization last Fall. This
reinstates it. (Running longer test of this now.)

Later, I could do away with AUXILIARY_OX_RADF directive and just turn this
on whenever clim_interact_chem=0, but wanted to minimize code changes for now.