Option to use an alternative gravity wave drag: Unresolved Wave Drag (Tiehan Zhou) This model uses all E4F40 model physics and horizontal and vertical resolutions except for our removing the existing Rayleigh drag and gravity wave treatments in that model.  Instead, we have used the orographic gravity wave treatment of McFarlane (1987), where the parameter Fr^2 = 0.5, and EKE is taken to be equal to 0.5 m^(-1).  This is applied by calculating orography variances over 2 x 2.5 GISS grid resolution using the high resolution topography from website of NGDC. The non-orographic gravity wave parameterization of Alexander and Dunkerton (1999) is also used.  We have used a B2 spectral source function (e. g., Gong et al., 2008) with time and latitude modulation to mimic the annual variation of tropical convection and extratropical storm tracks.  This forcing is applied at the model level immediate above 100 hPa.  We have used Cw (the spectral width) = 10m/s and Bm (amplitude) = 0.2 m^2/s^2 in this parameterization.  Note that the value of the wind at the source level (100 hPa) is also used at each gridpoint.
Both gravity wave parameterizations are applied at every physics step (every 30
minutes).  We run this model with a time step of 180 seconds.  This time step is
governed by the maximum winds in the model.  These occur at, or near, the top of
the model, and are sensitive mainly to the non-orographic gravity wave
parameterization settings.

Our settings give reasonable zonal mean wind and temperatures, but we have NOT
run extensive diagnostics on this model at this point, and the parameterization
settings cannot be optimized until this is done.