1. Input file Ginoux2001_source_VegMask_144x90_C90 for prefDustSources=0
2. Input file Ginoux_source_v2009_VegMask_0.25x0.25_to_C90 for
prefDustSources=1, which is the default and identical to
Ginoux_2001_updated_hires_withVegMask_v2009_C90. I removed "2001" from
the name to prevent a mix up with the old version of the preferred
dust source file, and I added the information that the data were
regridded from the file with 0.25x0.25 deg. horizontal resolution.