Nearly complete implementation of Parser and Dictionary (AssociativeArray) classes.  These implementations extend/generalize the treatment of rundeck parameters to allow multiple sets of parameters to exist simultaneously.  (E.g. each tracer has its own set of property/parameter values.)
A significant amount of documentation needs to be updated, and I have
not yet severed the ties to the original implementation.  Need to also
think some more about some of the primary names.  E.g. is "Dictionary"
better than "Map" or "AssociativeArray"?  Or perhaps "ParameterSet".
I can be easily swayed (other than "Map") if someone has a strong
opinion.