This is an initial attempt to merge various constants related to the Julian calendar from CONST.f and MODEL_COM.f.  There are many places throughout the model where these constants appear implicitly or as literals rather than properly referring to the symbolic names.
NOTE: I have not actually added any switches for this module to be
used yet.  We can gradually migrate files to use the new home for
these.  I will try to come back at some point in the near future and
"force" the issue.

For now, I have implemented two sets of names.  The legacy names are
identical to those in CONST and MODEL_COM.  The new names are more in
the spirit of the new coding standards document.  I will phase out the
legacy names once _after_ the remainder of the model switches to use
this file.

Also, note that there are 2 parameters which probably do not belong
here: EARTH_DAYS_PER_DAY (EDPERD) and EARTH_DAYS_PER_YEAR (EDPERY).
These vary depending on the planet and epoch, and perhaps belong
elsewhere.  Opinions?