Added a 3D array daily_z that stores the height of each level for use in the remapping/placement of 3D tracer emissions onto model levels.  So as not to cause problems with restarts after code updates, writes/reads of daily_z to/from the restart file are currently disabled.  They will need to be enabled once volcano and aircraft emissions begin to use daily_z.
The instantaneous model heights could be used for this purpose
(and already are in some cases), but restart reproducibility
would then require that the placement onto model levels be
done each timestep.  This is not desirable in all cases;
daily time resolution of level heights is a reasonable
compromise between efficiency and accuracy for current
vertical resolutions (of data and the model).