Introducing the parameter trname(n)//'_nBBsources' in the rundeck and corresponding usage by the model.
This parameter defines how many of the input files of tracers are biomass burning sources, in order to distribute them throughout the boundary layer, instead of only at surface. For now, this applies only to aerosols, although the rundeck parameters for gases are also committed, since very soon they will start being used. The usage of this parameter on aerosols removes the hardcoded number 2 that used to define the number of BB sources in the code. Still, BB sources need to be defined last in the emissions file list.