Following changes were for Shindell CH4 tracer sources (which are not active in the E4Tcad(i)F40 decks). Other tracers sources should not be affected.
(1) CH4 is added now to the list of tracers that are allowed to specify
sources that mix through the PBL with the *_nBBsources parameter settings.

(2) this necessitated streamlining the treatement of:
#ifdef WATER_MISC_GRND_CH4_SRC. Code is now a bit cleaner and
understandable, with respect to that directive. Trade-off is that
one more emission file must be listed when WATER_MISC_GRND_CH4_SRC
CPP is defined (this is a long story you don't want to know.)
Also, the CH4 Ocean, Lake, and "Misc Ground Source" for Shindell
CH4 case has now been combined into one source.

(3) changed a couple statements that would have caused trouble in very
rare model configurations like shindell tracers on but imAER==1.
But one was kinda important: moved this line:
call apply_tracer_3Dsource(nBiomass,n)
so that shindell tracers could still use PBL-mixing biomass sources
if TRACERS_AEROSOLS_Koch and TRACERS_AMP are off.

(4) Put in a model stop if #ifdef DYNAMIC_BIOMASS_BURNING
because the fire model emissions method is disabled pending an
overhaul soon.