(1) Currently (in TRACERS_DRV) there is a model stop if CPP  DYNAMIC_BIOMASS_BURNING is used. Protection while scheme is upgraded.
(2) This is an update to the flammability code to the section that does emissions
only (not the flammability itself). Rather than having emission-per-fire-count
defined on an per gridbox basis per tracer from an input file. Now (or when
TRACERS_DRV is update), emissions-per-fire-count for relevant tracers are PER
VEG-TYPE where the types are ye olde 12 GISS model types, or Ent remapped to them.

[Cubed sphere people: so there are no trname_EPFC files any more, which is why we
did not give you them for regridding.]

The idea here, of course, is that the emissions can now change when veg types change.
Unfortunately, the flammability is still dependant on a vegetation density input file;
once vegetation density is available for the online types, we can switch to this.

Ent/Veg-people: This type remapping is done in the same way it is currently done
in RADIATION_DRV (thanks Igor), see the line: temp hack: ent pfts->giss veg. And
"!YKIM-temp hack" there. Therefore as the vegetation model evolves so must the
fire model emissions. If Igor or YKIM could make a note of that dependency, that
would be great, but of course I will keep an eye on it.

(3) To match upcomming emissions-per-fire-count numbers that will be put in
TRACERS_DRV, changing the value of mfcc the MODIS fire count calibration
paramter.