** BUG FIX ** At some point, my *guess* is during the massive aerosols modernizing on 5/11/2010,  a call was put in to apply the 3D aircraft source for all tracers, not knowing that there was already a call to do this for NOx. Consequently aircraft NOx sources have been double the source file's intent. I am fixing this by making that call conditional with a if(n>ntm_chem). Kostas, please confirm this is OK with you. Also, please check other aircraft sources like SO2 that this double-counting is not there too.
So while it is important that tracer people update, this shouldn't affect ongoing balancing efforts since no aircraft in 1850.

Similarly, at the same place in the code there is an updating of 3D biomass burning for all tracers. To protect the option of #defined GFED_3D_BIOMASS, I am changing:

call apply_tracer_3Dsource(nBiomass,n)

to:

#ifndef GFED_3D_BIOMASS
        call apply_tracer_3Dsource(nBiomass,n)
#else
        if(n>ntm_chem)call apply_tracer_3Dsource(nBiomass,n)
#endif /* note that is a NOT defined */

Side note: the aircraft bug was easy to miss because it didn't show up in setup hour because the first call applied zero source, since the source had not been defined yet.