This commit changes online fire emissions when DYNAMIC_BIOMASS_BURNING  is on (gasses and aerosols). [Also removes the temporary model stop that was in place.] Shouldn't change sources otherwise.  It goes along with the commit to flammability.f from early June, so I paste a summary that below. But first, the specifics to this commit:
No more EPFC (emissions per fire count) files. The do_fire parameter is now set to be true for all tracers that allow fire model emissions whenever DYNAMIC_BIOMASS_BURNING CPP is defined. So for now, it is all or nothing. These are: CH4, NOx, CO, Alkenes, Paraffin, BCB, OCB, NH3, SO2 (and thereby SO4). Each of the above tracers has a section that defined emisPerFireByVegType(n,1:12) in the section e.g. where tr_mm(n) is defined. [This will be something that would be put in some external file once Tom's modernization takes place.] These numbers from Olga Pechony are now consistent with the MODIS fire count calibration
paramter change in flammability.f and chosen to be consistent with the AR5 2000 grassfire+forestfire sources.

If you were using the fire model with non-Koch/Shindell tracers (like AMP) please check that it is working.

The fire model emissions are no longer 2D surface but mixed throughout the PBL. Model stop put in if  do_fire is true and nBBsources > 0, which would be inconsistent. [I also put in a couple checks to make sure e.g. ntsurfsrc(n)+nBBsources(n) is not > ntsurfsrcmax.]

[The special case of methane (Shindell not Lerner) has been made slightly simpler by combining the three online source when WATER_MISC_GRND_CH4_SRC is on into one and placing in a slot .le. ntsurfsrc(n). So no longer need special attention to this when doing fire model or nBBsources sources. This needs the currently commented-out file: CH4_##=AR5_emis/F/NAT/CH4gsfMGOLjal_blank_2x2.5_h]

Kostas' bb_fact and src_fact for Aerosol biomass burning remain at 1.0 if do_fire is true.

Now, edited notes follow from the commit notes to flammability.f from earlier in the month:

...update to the flammability code to the section that does emissions only (not the flammability itself). Rather than having emission-per-fire-count defined on an per gridbox basis per tracer from an input file, emissions-per-fire-count for relevant tracers are PER VEG-TYPE where the types are ye olde 12 GISS model types, or Ent remapped to them.

The idea here, of course, is that the emissions can now change when veg types change. Unfortunately, the flammability is still dependant on a vegetation density input file; once vegetation density is available for the online types, we can switch to this. Until then, when modeling a different era, consider swapping this file.

Ent/Veg-people: The veg type remapping is done in the same way it is currently done in RADIATION_DRV (thanks Igor), see the line: temp hack: ent pfts->giss veg. And "!YKIM-temp hack" there. Therefore as the vegetation model evolves so must the fire model emissions. If Igor or YKIM could make a note of that dependency, that would be great, but of course I will keep an eye on it.