Adding a stomatal dry deposition diag for Ox tracer by defining new PBL variable for the stomatal deposition velocity which is calculated in TRDRYDEP.f, very similar to the total deposition velocity.
The diag itself is only accumulated/written for #defined ACCMIP_LIKE_DIAGS for the moment. And only over ice-free land.

We are actually not yet sure if this diag is "correct." It just seemed like a logical way of doing this, and passes the most basic tests like being less than the total deposition and being a larger fraction of the total over highly vegetated areas.
I am assuming it would be better to use something related to the stomatal conductance that appears in VEG_DRV: However, since the Ox total dry deposition still comes from TRDRYDEP.f, what we have done here is more consistant.

TRDIAG_COM.f: Increasing ktajls (not related to current changes here, just because I tested with ACCMIP_LIKE_DIAGS on.)