Except for a limited testing period, warnings should be written to unit 0 only if followed by a call to stop_model.
Calls to stop_model due to unphysical circumstances now write statements
starting with

Model crashed due to ...

to unit 0 . This may be used to "qdel" a job that does not abort if this
occurs. Hopefully, intel will soon have a fix for that compiler bug and the
abort will be automatic.