1. New parameter n_soilDust for the first soil/mineral dust tracer index. It is assigned in subroutine init_dust in TRDUST_DRV.f only at initialization time. Like it is done in DIAG.f now, it will successively replace all the individual assignments of the first index, currently spread all over the code and done every time step.
2. Parameter dust_names assigned in subroutine init_dust in
TRDUST_DRV.f from now on.

3. To get the new netcdf i/o to properly work for subdd,
subdd-variables TMIN and TMAX, which are written only at the end of
the day, need to be treated separately from the other variables that
are written every NSUBDD time step.

4. New subdd-variable DuCONC for the three-dimensional concentration
of soil dust aerosols.

5. Removed obsolete subroutine write_2d_tracer. New subroutine write_4d
to write four-dimensional variables as netcdf in subdd (very similar
to write_3d, just with one more dimension).

6. Removed some redundant code. Code bits.