After discussions with Gavin S. and Rick H. today, committing changes such that the template decks that use chemistry_params and chemistry_input_files will be set to have the fastj2 photolysis code's solar flux change with year and be consistent with rad code input file: solar.DBbglean.ann850-2000.uvflux_hd
The year used follows s0_yr if non-zero and jyear if s0_yr=0,
IF YOU ALSO update TRCHEM_init.f to be committed momentarily. This is
important because otherwise it will just follow jyear.

In chemistry_params this amounts to setting rad_FL=1
In chemistry_input_files this amounts to listing the file:
solar.DBbglean.ann850-2000.uvflux_hd_fastj2