Multidecadal coupled simulations using the cubed-atmosphere configuration have revealed that runaway in-situ ice production can occur at coastal gridcells (both with Ocean R and HYCOM).  Moving sea ice processes to the ocean grid would help prevent this, but until then, an alternative is necessary.
Existing safeguards against advective accumulation of coastal ice
have been replaced by an actual expulsion mechanism.  A corrective
transport velocity is added to the velocities predicted by the ice
dynamics.  It is nonzero only where the ocean fraction is < 100% in
one of two neighboring cells, and where ice thickness is larger in
the cell having the smaller ocean fraction.  The magnitude of the
corrective velocity is proportional to the ice thickness gradient.