Added a variant of write_dist_data based on nonblocking I/O procedures in the latest release of parallel-netcdf. The new functionality takes advantage of the available economies of scale writing large chunks of data to parallel filesystems.  It will initially be deployed in modelE to write tracer arrays to checkpoint files and to write collections of diagnostics to "subdd" files. I/O of other fields may be batched up later.
The new subroutines in the pario module write collections of
netcdf variables of rank 2, 3, or 4, which are passed in as
arrays of rank 3, 4, or 5 together with an array of netcdf names.
So, each collection currently must be comprised of arrays of the
same rank.

As versions 3.x of Intel MPI sometimes mishandle the case when a
batched-up write request corresponds to multiple regions within
a file, the netcdf variable IDs in each request must be consecutive.
I've added a consecutiveness check and a stop if necessary;
the problem does not occur with Intel MPI 4.x but there are
performance implications anyway.