Added coordinate variables as well as units and long names of variables and some global attributes to high frequency netcdf output files. I have tried to be as COARDS compliant as possible with the coordinates.
Longitude and latitude coordinates are in Degrees East and Degrees
North, respectively. The vertical coordinate for three-dimensional
arrays is either pressure level, sigma level, or ground layer,
depending on the physical variable.

Each file contains three time variables, which are internal time unit
(ITU) "itime", "time", and "calendar".

ITU is variable "itime" in modelE, which counts the number of time
units of length Dtsrc since the reference date 01/01/iyear1.

The time variable "time" is the actual COARDS compliant time
coordinate that is recognized by various applications (with some
caveat, see below). "time" counts hours or days relative to a
reference date or time. The reference date or time is either the
midpoint of the last ITU of the first nsubdd ITUs after 01/01/iyear1
for instantaneous variables, or the midpoint of the first nsubdd ITUs
for accumulated or time averaged variables, or 01/01/iyear1 for daily
variables. Accordingly, the current counted hour or day of "time" in
an output file is either the midpoint of the last ITU of nsubdd ITUs,
or the midpoint of nsubdd ITUs, or the current count of the day,
depending on the type of the output.

"calendar" is a string variable that contains the current modelE date
and time, or only the date for daily variables, to make it easier to
relate the model calendar to variable "time".

I added the attribute "relation_to_itime" to the time variables "time"
and "calendar" as a reminder how they are related to "itime".

The global attribute "xlabel" prints the variable xlabel from the
simulation. The global attribute "calculation_method" indicates
whether the data are "instantaneous value for current ITU",
"accumulated/averaged over nsubdd ITUs", or "daily". Finally, the
global attribute "nsubdd" shows the setting of the nsubdd-parameter.

I already have added units and long names for most variables in
get_subdd, except for a few derivates of Be7-tracers, some obscure
cloud variables, and the ones I simply missed.

Three-dimensional variables, the vertical dimension of which is on
sigma levels, or on ground levels are written as three-dimensional
arrays. Nsubdd-variables on pressure levels with the names "ALL" in it
are written as three-dimensional arrays as well, if write_subdd is
called with both the optional argument "positive=up" and the optional
argument "suffixes" and the argument-value is an array of strings with
the pressure levels. If only "suffixes" is present, those variables
are written as two-dimensional arrays with a separate name for each
level. Latter is an option, but hasn't been applied for any variable,
currently.

I tested the netcdf-writeout for various rundecks using model
configurations with the traditional lat-lon grid, although I mainly
paid attention that it works technically. I haven't checked each
variable whether the output actually makes sense from a physical point
of view. Everyone who uses this should check to find possible
glitches.

Since I still have to figure out how the netcdf-output for subdd works
with the cubed sphere version of the model, and to avoid any
compilation or runtime problems, the new i/o for subdd is set to
undefined at the top of DIAG.f, when the cubed sphere is
defined. Thus, the new i/o for subdd doesn't work for the cubed sphere
model yet.

Note, regarding applications for netcdf-files. Various viewing
applications, like GrADS or Panoply require the dimension names lon,
lat, and time for the longitude, latitude, and time coordinate,
respectively. The model's dimension names are im, jm, and record
instead. This can be easily resolved by running the netcdf-file
through ncrename, which is one of the nco-programs developed by
Charlie Zender. This looks like

nco -d im,lon -d jm,lat -d record,time <file.nc>

or see documentation for nco.

In addition, the COARDS-standard requires the attribute "calendar" for
the time-coordinate, which is set to "noleap", since modelE doesn't
know leap years. However, recent versions of GrADS don't accept
"noleap" in self-describing files anymore. This can be solved by
calling the netcdf-file using a descriptor file that contains the
description for the time-dimension.

Additional edits: relative humidity in % and daily maximum and minimum
temperatures in deg. Celsius.