(1) F0( ) for Ox tracer increased back to 1.4 from 1.0.
(2) fixing a very old bug where a factor of 1.e-3 was missing from ssfac0 in both the condensation of gas tracers into cloud water and into precipitation in washout. Thanks to Yunha Lee for noticing.

(3) Applying gasPrecFracScal: scaling factor on the effective precipitating fraction for gases for below-cloud washout. This is applied when thlaw is calculated only. Defaults to 1, but is set to 15 in the E4TcadF40 and E4TcadiF40 template decks for now.

Note that for both (2) and (3) above I only changed GET_COND_FACTOR_array and GET_WASH_FACTOR_array. I did NOT change GET_COND_FACTOR or GET_WASH_FACTOR, not knowing if those were supposed to be maintained as-is. Please change them if they need to be, or ask me to.

If you want more information on these and I have not just e-mailed you after committing this, please let me know: gfaluvegi@giss.nasa.gov