Updated a number of rundecks that are used for nightly regression tests.  First and foremost, these were modified to use NEW_IO which overcomes some annoying technical problems with using CMPE002 in some cases.
Also, I had to update several rundecks to use the new RADN7, and to
enforce the policy that setup runs just 1 hour, and the continuation
runs 1 day.

Finally, in a 2 instances I had to modify ISTART to be 2 to bootstrap
to allow netcdf.  (Appropriate files were not available for "warm"
startup.)