Added the NEW_IO analog to CMPEOO2P, to allow certain diagnostic arrays to be omitted from reproducibility checks:
(1) diffreport accepts an optional command-line argument indicating
which variable's attributes comprise a list of variable names to skip

(2) in the model, subroutine def_rsf_acc() defines a variable
is_npes_reproducible in checkpoint files, and populates its
attributes with the list of accumulation arrays known to have
roundoff differences on differing processor counts

The attribute is_npes_reproducible could have been defined for the
relevant arrays; I may switch to that alternative if it has some
advantage over (1).  The current choice allows grouping of all the
info in one place in the checkpoint file.

The simple on/off logic could eventually be replaced by specification
of maximum acceptable relative/absolute differences.  However, this
is trickier than it sounds, in part due to the fact that a given
acc-array contains fields having different orders of magnitude and
relative susceptibilities to roundoff issues.