There was a bug in the interpolation of transient monthly tracer sources between time slices in the emissions files. Unfortunately, I recommend updating to this fix for the transient emissions portions of AR5 runs (frozen model?).
Fortunately, it's my understanding that this does not affect 1850
runs that have been done (e.g. for balancing/PI spinup), because I
believe those used O3_yr=-1850 and aer_int_yr=1850 (that is, non-
transient emissions were used, even though transient emissions
files are listed in the decks).

I have tested this fix by running setup multiple times and checking
the interpolation. When current discover problem is fixed we're
going to start some annually-varying emissions runs that are
sensitive to this bug, so these will serve as a longer test.

About the bug (or what happens if you don't update): Depending on
time of year, there was unintended skewing of the linear
interpolation in time for monthly sources. For emissions files that
vary by decade (like the AR5 forest fire, grass fire, shipping and
aircraft files) in transient mode (o3_yr=0 and/or aer_int_yr=0):
I think the error maxes-out at 5%. E.g. at the start of a year
you'd be using about 5% too much contribution from the earlier
decade, and at the end of a year 5% too much contribution from the
later decade. Correct answer mid-year. The 5% here is the ratio of
half a year to the 10 year time-slice step. Therefore, the problem
gets worse if you use higher time resolution slices.

(Note that after the fix, monthly emissions files that vary from one
year to the next are in effect no longer interpolated between years,
as it should be.)