Temporary fix so that E001tr can pass the regression tests.
Revision 2.265 of TRACERS_DRV.f moved the South African
SF6 source from the continental interior to the coast;
with the no-fractional-ocean 4x5 TOPO file used in E001tr,
the source is in a single all-ocean gridbox.
I removed the multiplication by fearth so that the total
area of the source is greater than zero.