Bug fix, following update to AR5 branch:
With use_rad_ch4=1 and ghy_yr=0, the L=1 overwriting of tracer CH4 with rad code updating values was not happening. Instead, the initial conditions were set correctly from the rad code, and the L=1 tracer value just maintained (so in effect, maintaining the initial conditions from the rad code at L=1). Obviously a problem for long transient runs. This should fix.

This version has a few extra notes in it: because the top-of-model ozone overwriting is doing the same thing, but we would like to leave it like that for the moment. Some comment lines were put in to show were we might change that.