This is the first of a series of pervasive changes in the way that modelE manages tracer metadata such as molecular mass, decay rate, etc.  Tracer metadata is now managed by an associative array (also called a hash or a dictionary).  For now, this aspect is still somewhat hidden in the implementation, and a set of tailor functions and subroutines are provided for accessing.  This reduced risk during the transition and greatly reduced the number of changes in the various parameterizations.  It will also aid in creating an data file which will replace the hardcoded values currently in the code.
Some simple examples of the changes:

Before:  TRW0 is an array sized NTM
              TRLAKE(:,1,I,J)=MLK1*TRW0(:)
After:   TRW0 is a function which returns an array
              TRLAKE(:,1,I,J)=MLK1*TRW0()

Before:  tr_mm, trdecay are arrays
              taij(i,j,k)=taij(i,j,k)*trdecay(n_Be7)*tr_mm(n_Pb210)
After:   tr_mm, trdecay are functions (subtle eh?)
              taij(i,j,k)=taij(i,j,k)*trdecay(n_Be7)*tr_mm(n_Pb210)


Soon these accessor procedures will be replaced by a more general
access procedure. E.g.

      taij(i,j,k)=taij(i,j,k)*getProperty(tracers,'n_Be7','trdecay')* &
     &     getProperty(tracers,'Pb210','tr_mm')

or when we updgrade to F2003, the slightly cleaner:

      taij(i,j,k)=taij(i,j,k)*tracers%getProperty('n_Be7','trdecay')* &
     & tracers%getProperty('Pb210','tr_mm')

The properties will also soon be getting clearer names such as
"decayRate" and "molecularMass".  And we will probably move teh
accesses outside of the inner loops both for clarity and for
performance.

If you need to add a new type of metadata before these transitions are
completed please contact me.  The "new" way will work already but
we'll need to coordinate.

NOTE: This is only tested with the E001tr rundeck, so I may have
missed some uses of tracer metadata not active with that rundeck.
Please test against your own favorite rundeck.